Extract ClientSocket.Server partial class
